Add manual dispatch button and fix tags to be latest
The dispatch button can be useful in some cases, to force a rebuild...

And the tags where being created with the branch name (main), when
what we want is the default "latest" one (and tags when there are
releases (that's correct).

In any case, this GHA is missing some little details like:

- Have a test phase to verify it works before building.
- Also run on PRs (not only on push).
- Only build the images when the repository is the moodlehq one.

That's not being fixed here, but surely should be considered.